Replace external 'int rawspeed_get_number_of_processor_cores()'
The requirement to have an external symbol that needs to be provided by
the executable just to set some number is way more complex than
necessary, especially as there is a sane default value available.

This removes this requirement and adds a simple `setThreadCount`
function to be used if the client code actually needs to change the
default.

This removes one of the ABI changes between builds with and without
NO_PTHREAD. Should also help in simplifying a shared library setup.