RawImage: replace manual pthread_mutex handling with ThreadSafeVector
Replace the two remaining mutex members by incorporating the locking
into the ThreadSafeVector<T> class. Thereby switching from a verbatim
pthread_mutex_t member to a pimpl idiom to achieve the main goal of
having a stable public ABI of the RawImageData class, independent of
wether or not NO_PTHREAD is defined.

This also makes the code safer in the sense that it is impossible to
modify the underlying vetor without locking.

The new class is not completely thread safe (see begin/end) but it is
strictly safer than the old code used to be and sufficiently safe for
the current use cases.